likhAI/requirements.txt
Flask==2.1.0
pytesseract==0.3.8
Pillow==9.0.0
transformers==4.18.0
torch==1.11.0
nltk==3.6.7
sentencepiece==0.1.96
protobuf==3.20.0
sacremoses==0.0.53